refactor(ui): decompose chat panel and sidebar for EVE-123
