+++ apps/ui/src/hooks/index.ts
@@ -17,3 +17,4 @@ export * from "./use-notifications";
 export * from "./use-mount-effect";
 export * from "./use-scroll-manager";
 export * from "./use-image-drop-zone";
+export * from "./use-chat-model-selection";
